backend/requirements.txt
# DESTINEX Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
pymongo==4.6.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0